backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
pydantic-settings==2.6.1
sqlalchemy==2.0.36
asyncpg==0.30.0
redis==6.4.0
httpx==0.28.1
alembic==1.16.4
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.0
email-validator==2.3.0
structlog==24.4.0
celery==5.6.0
boto3==1.35.36
python-multipart==0.0.22
prometheus-client==0.22.1
slowapi==0.1.9
stripe>=10.0.0
sentry-sdk[fastapi]>=2.0.0
resend>=2.0.0
